resolved circular dependencies of account preferences by extracting enums, classes, and typedefs into a dedicated account_pref_types leaf module, allowing account_repository to import it instead of the full preferences controller
